requirements.txt
numpy
faiss-cpu
sentence-transformers
beautifulsoup4
requests
